Moved html-to-view helpers to shared helper class for reuse between popup viewer and feature form view.
Also added support for the `<code>` tag
